+++ app/controllers/admin/all_galleries.py
@@ -1,10 +1,9 @@
-from app.allImports import * 
+from . import admin
 from app.logic.validation import *
-from werkzeug.security import check_password_hash
-from flask import session
 from app.models.Forms import *
 
-@app.route('/', methods=["GET","POST"])
+from flask import render_template
+@admin.route('/', methods=["GET","POST"])
 def all_galleries():
     user = Forms.get(fid=1)
     name = user.first_name
