+++ app/controllers/public/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+public = Blueprint('public', __name__)
+
+from . import application_create
+from . import application_review
