+++ app/templates/views/gallery_view.html
