+++ scripts/dummy_data.py
@@ -22,7 +22,7 @@ def create_tables():
         model.create_table()
 
 def dummy_data():
-    drop_tables()
+    # drop_tables()
     create_tables()
     load_user_data()
     load_files_data()
